src/ingestion/layoutlm_analyzer.py
import torch
from transformers import LayoutLMv3Processor, LayoutLMv3ForTokenClassification
from PIL import Image
import logging

logger = logging.getLogger(__name__)

class LayoutLMAnalyzer:
    def __init__(self, config: dict):
        self.config = config
        self.device = torch.device(config["ingestion"]["device"] if torch.cuda.is_available() else "cpu")
        self.model_name = config["ingestion"]["model_name"]
        self.processor = LayoutLMv3Processor.from_pretrained(self.model_name, apply_ocr=False)
        self.model = LayoutLMv3ForTokenClassification.from_pretrained(self.model_name).to(self.device)
        self.confidence_threshold = config["ingestion"]["confidence_threshold"]

    def analyze(self, image_path: str) -> dict:
        """
        Analyze screenshot with LayoutLMv3 to extract:
        - Text (English + Korean)
        - UI elements (red dots, arrows, decision diamonds, dimmed icons, toast popups)
        - Gestures (swipe down, tap, etc.)
        """
        image = Image.open(image_path).convert("RGB")
        width, height = image.size

        # LayoutLM expects image + bounding boxes — we’ll use OCR-free mode and detect elements via post-processing
        # In OCR-free mode, we need to provide empty text and empty boxes to avoid the 'words' and 'boxes' key errors
        encoding = self.processor(image, text=[""], boxes=[[]], return_tensors="pt").to(self.device)
        with torch.no_grad():
            outputs = self.model(**encoding)

        # Post-process outputs to detect UI elements
        # In real implementation, this would use LayoutLM’s token classification + bounding box regression
        # For now, simulate with dummy data — replace with real LayoutLM logic later
        detected_elements = self._detect_ui_elements(image, outputs, width, height)

        return {
            "text": self._extract_text(encoding, outputs),
            "ui_elements": detected_elements,
            "image_path": str(image_path),
            "width": width,
            "height": height
        }

    def _detect_ui_elements(self, image, outputs, width, height):
        # Simulated UI element detection
        # In real code, use LayoutLM’s token classification + bounding box to detect:
        # - Red dots + arrows → gesture
        # - Decision diamonds → conditions
        # - Dimmed icons → disabled state
        # - Toast popups → error messages
        return [
            {
                "type": "gesture",
                "subtype": "swipe_down",
                "target": "shutter_button",
                "bbox": [100, 200, 150, 250],
                "confidence": 0.95
            },
            {
                "type": "condition",
                "name": "timer_enabled",
                "bbox": [300, 400, 350, 450],
                "confidence": 0.85
            },
            {
                "type": "error",
                "message": "Battery too low to use flash.",
                "bbox": [500, 600, 600, 650],
                "confidence": 0.90
            }
        ]

    def _extract_text(self, encoding, outputs):
        # Simulated text extraction
        # In real code, use LayoutLM’s OCR + token classification
        return [
            {"text": "Flash Off & Dim condition", "lang": "en", "bbox": [500, 600, 600, 650]},
            {"text": "플래시 꺼짐 및 비활성 상태", "lang": "ko", "bbox": [500, 600, 600, 650]}
        ]